requirements.txt
Pillow>=9.5.0
requests>=2.28.2
numpy>=1.23.5